Before: 5472521583 ns/op (5.47s)
After: 3350061625 ns/op (3.35s)

Execution time decreased by about 38.8%

```plain
pkg: github.com/gnoswap-labs/tlin/internal
BenchmarkEngine-8            	       1	5472521583 ns/op	20200576 B/op	  168915 allocs/op
BenchmarkGoroutineEngine-8   	       1	3350061625 ns/op	19007344 B/op	  168759 allocs/op
```